chore: Add pagination functionality to search results table
This commit adds pagination functionality to the search results table in the Principal component. The Pagination component is imported and used to display the paginated data based on the selected number of records per page. The data is filtered based on the search parameter and the current page number. The pagination buttons allow the user to navigate to the previous and next pages, as well as select a specific page. The number of records per page can also be adjusted using the dropdown menu.

```